Rename file containing JANI expression support functionalities
