+++ src/as2fm/jani_generator/jani_entries/jani_assignment.py
@@ -20,7 +20,7 @@ Assignment in Jani
 from typing import Dict
 
 from as2fm.jani_generator.jani_entries import JaniConstant, JaniExpression, generate_jani_expression
-from as2fm.jani_generator.jani_entries.jani_convince_expression_expansion import expand_expression
+from as2fm.jani_generator.jani_entries.jani_expression_support import expand_expression
 
 
 class JaniAssignment:
